MERGE: trunk revisions r5500:5600
git-svn-id: svn://scm.gforge.inria.fr/svnroot/sofa/branches/Sofa-1.0@31 b7f1a160-662a-0410-a6d8-bea165342f16

+++ modules/sofa/simulation/common/Visitor.cpp
@@ -152,7 +152,7 @@ void Visitor::printInfo(const core::objectmodel::BaseContext* context, bool dirD
 
   double Visitor::getTimeSpent(ctime_t initTime, ctime_t endTime)
   {
-    return endTime-initTime;
+    return (double)(endTime-initTime);
   }
 
   void Visitor::printNode(const std::string &type, const std::string &name, const TRACE_ARGUMENT &arguments)
